extract item view and image to separate structs
